+++ changelog.md
@@ -1,1 +1,2 @@
 - Updated NeoForge
+- Fixed image items not being able to load